Publish a developer package to GitHub Packages (#7)
* feat(gha): publish developer package

Introduces a GitHub Actions workflow `publish-dev-package` which
publishes a developer package to the GitHub npm registry when commits
are pushed to the `main` branch. A developer package bears the target
release version but followed by the short commit hash of the commit used
to build the package. The actual steps are described in
`publish-package.yml`.

* docs: update README

Topics:
- Explains how to install a developer package from GitHub Packages
- Explains how to configure a GitHub personal access token to install a
  developer package